refact(auth-user-onboarding) : Remove barrel imports for inputField component
